backend/pyproject.toml
[project]
name = "valerie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "prefect>=3.0.0",
    "httpx>=0.27.2",
    "python-dotenv>=1.0.1",
    "pymongo>=4.8.0",
    "boto3>=1.35.0",
    "openai>=1.0.0",
]
